Use the new `favor` option when measuring queries
The Ra system is also configured with a WAL size of 64 MiB, like the Ra
system used in RabbitMQ for the integration of Khepri. The impact is
mostly a smaller memory footprint.

The code was refactored a lot to allow different "profiles" (rather than
`safe` and `unsafe` for everything).

Colors are updated as well in the process:
* inserts and deletes are using green
* queries are using blue

As a consequence, we don't measure queries with fsync(2) disabled
anymore. It didn't make sense anyway, we measured the same thing twice
between `safe` and `unsafe` because the disk was not involved.